Always use error logger instance from `GiniConfiguration`
This makes sure our error logger is called also when using the
component api.

Also using `String(describe: error)` instead of
`error.localizedDescription` because only for `NSErrors` and swift
errors conforming to `LocalizedError` does it return a meaningful
message.

Fixed an async issue in `GiniScreenAPICoordinate+Camera` which
caused the `error.localizedDescription` to be always logged.